changed size argument type from gulong to gsize as discussed on
2008-01-29 14:58:31  Tim Janik  <timj@imendio.com>

        * glib/gmem.[hc]: changed size argument type from gulong to gsize as
        discussed on gtk-devel-list:
          http://mail.gnome.org/archives/gtk-devel-list/2007-March/msg00062.html
        this should be ABI compatible on all platforms except win64 for which
        no ABI binding port exists yet.



svn path=/trunk/; revision=6413

+++ glib/gmem.h
@@ -44,15 +44,15 @@ typedef struct _GMemVTable GMemVTable;
 
 /* Memory allocation functions
  */
-gpointer g_malloc         (gulong	 n_bytes) G_GNUC_MALLOC;
-gpointer g_malloc0        (gulong	 n_bytes) G_GNUC_MALLOC;
+gpointer g_malloc         (gsize	 n_bytes) G_GNUC_MALLOC;
+gpointer g_malloc0        (gsize	 n_bytes) G_GNUC_MALLOC;
 gpointer g_realloc        (gpointer	 mem,
-			   gulong	 n_bytes) G_GNUC_WARN_UNUSED_RESULT;
+			   gsize	 n_bytes) G_GNUC_WARN_UNUSED_RESULT;
 void	 g_free	          (gpointer	 mem);
-gpointer g_try_malloc     (gulong	 n_bytes) G_GNUC_MALLOC;
-gpointer g_try_malloc0    (gulong	 n_bytes) G_GNUC_MALLOC;
+gpointer g_try_malloc     (gsize	 n_bytes) G_GNUC_MALLOC;
+gpointer g_try_malloc0    (gsize	 n_bytes) G_GNUC_MALLOC;
 gpointer g_try_realloc    (gpointer	 mem,
-			   gulong	 n_bytes) G_GNUC_WARN_UNUSED_RESULT;
+			   gsize	 n_bytes) G_GNUC_WARN_UNUSED_RESULT;
 
 
 /* Convenience memory allocators
@@ -123,7 +123,7 @@ typedef struct _GMemChunk  GMemChunk;
 #define G_ALLOC_AND_FREE  2
 GMemChunk* g_mem_chunk_new     (const gchar *name,
 				gint         atom_size,
-				gulong       area_size,
+				gsize        area_size,
 				gint         type);
 void       g_mem_chunk_destroy (GMemChunk   *mem_chunk);
 gpointer   g_mem_chunk_alloc   (GMemChunk   *mem_chunk);
